create a new ant task to generate skeleton
