+++ nbbuild/build.xml
@@ -1591,13 +1591,17 @@ It is possible to use -Ddebug.port=3234 -Ddebug.pause=y to start the system in d
     <target name="verify-all-libs-and-licenses" depends="bootstrap,download-all-extbins" description="Verify the contents of third-party libraries and licenses.">
         <taskdef name="verifylibsandlicenses" classname="org.netbeans.nbbuild.extlibs.VerifyLibsAndLicenses" classpath="${nbantext.jar}"/>
         <property name="verify-libs-and-licenses.haltonfailure" value="false" />
-        <verifylibsandlicenses report="${nb.build.dir}/verifylibsandlicenses.xml" mavenfolder="${nb.build.dir}/mavenpoms" nball=".." haltonfailure="${verify-libs-and-licenses.haltonfailure}" />
+        <verifylibsandlicenses report="${nb.build.dir}/verifylibsandlicenses.xml" nball=".." haltonfailure="${verify-libs-and-licenses.haltonfailure}" />
+        <taskdef name="mavenskeletonbuild" classname="org.netbeans.nbbuild.extlibs.MavenSkeletonProject" classpath="${nbantext.jar}"/>
+        <mavenskeletonbuild mavenfolder="${nb.build.dir}/mavenpoms" nball=".." />        
     </target>
 
     <target name="verify-libs-and-licenses" depends="bootstrap,download-selected-extbins" description="Verify the contents of third-party libraries and licenses in the selected configuration.">
         <taskdef name="verifylibsandlicenses" classname="org.netbeans.nbbuild.extlibs.VerifyLibsAndLicenses" classpath="${nbantext.jar}"/>
         <property name="verify-libs-and-licenses.haltonfailure" value="false" />
-        <verifylibsandlicenses report="${nb.build.dir}/verifylibsandlicenses.xml" mavenfolder="${nb.build.dir}/mavenpoms" nball=".." haltonfailure="${verify-libs-and-licenses.haltonfailure}" />
+        <verifylibsandlicenses report="${nb.build.dir}/verifylibsandlicenses.xml" nball=".." haltonfailure="${verify-libs-and-licenses.haltonfailure}" />
+        <taskdef name="mavenskeletonbuild" classname="org.netbeans.nbbuild.extlibs.MavenSkeletonProject" classpath="${nbantext.jar}"/>
+        <mavenskeletonbuild mavenfolder="${nb.build.dir}/mavenpoms" nball=".." />        
     </target>
 
     <target name="create-mandatory-files-for-binary" depends="bootstrap,download-selected-extbins" description="Create a summary of the licenses used by libraries in the build.">
